Fix jQuery/bootstrap.min.js/affix/ocaml.org embedding left sidebar bug
When jQuery and bootstrap.min.js are loaded, the left side menus in
Documentation and Blog do not exert appropriate right pressure on the
main span9 content. This causes the content to jump (after the js load)
to overtop the left menu. This makes the page unusable.

After this change, the side menu no longer scrolls with the page
(position does not get changed to fixed by JS). It is possible to get
some of the previous functionality back by adding a height of >=25px to
the span/div first child of the span3 left sidebar element. This still
doesn't quite fix the issue as the menu elements then sometimes slightly
overlap the content. This patch disables all of the insane JS
modification of DOM element classes to set the menu position:
fixed. There is probably a way to get the effect we want without using
JavaScript. It would be nice to not have to load any JavaScript for a
simple static site...

Right now JavaScript is required for some of the package index search
features and the Documentation dropdown menu for some reason. The
package search should have a server-side fallback and the dropdown
should be achievable in pure CSS.

The gauntlet is laid! It is now your mission, young wanderer, to
discover what riddles wait in...

THE LAND OF WEB PAGE LAYOUT MAGIC

*fade to black*