Update Dockerfile to us Alpine 3.20.
Update the `build-opam2web` step to use opam 2.1 as per the other steps.
The performance improvements in opam 2.1 over 2.0 mean that the timeout
added as a hotfix in #237 is no longer needed.

Update Dockerfile to Caddy 2.8.4 (which is built on Alpine 3.20).
It may be worth considering using `caddy:latest` to address emergent
vulnerabilities.